feat: add Chief Maintenance Engineer (CME) role and update related functionality

+++ migrations/add-user-ba-no.sql
@@ -1,2 +1,2 @@
--- Add user_ba_no column to users table
-ALTER TABLE users ADD COLUMN IF NOT EXISTS user_ba_no INTEGER UNIQUE;
+-- Add user_ba_no column to users table (use BIGINT to support large BA numbers)
+ALTER TABLE users ADD COLUMN IF NOT EXISTS user_ba_no BIGINT UNIQUE;
